`get_thread_name_from_type`を`thread_type.py`に移動
